Refine usage of np.ndarray
- Add missing dtype params.

+++ nautilus_trader/indicators/average/sma.pyx
@@ -109,7 +109,7 @@ cdef class SimpleMovingAverage(MovingAverage):
         self._increment_count()
         self._inputs.append(value)
 
-        self.value = fast_mean(np.asarray(self._inputs))
+        self.value = fast_mean(np.asarray(self._inputs, dtype=np.float64))
 
     cdef void _reset_ma(self) except *:
         self._inputs.clear()
